ENH: improve _count_paths_outside_method via pythran
